complete excercise 5(auth)

+++ src/types/product/product.resolvers.js
@@ -11,24 +11,34 @@ const productsTypeMatcher = {
 
 /** product */
 const product = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError('you are not authenticated')
+  }
   return Product.findById(args.id)
     .lean()
     .exec()
 }
 
 const newProduct = (_, args, ctx) => {
-  // use this fake ID for createdBy for now until we talk auth
-  const createdBy = mongoose.Types.ObjectId()
-  return Product.create({ ...args.input, createdBy })
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError("you don't have access")
+  }
+  return Product.create({ ...args.input, createdBy: ctx.user._id })
 }
 
 const products = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError('you are not authenticated')
+  }
   return Product.find({})
     .lean()
     .exec()
 }
 
 const updateProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError("you don't have access")
+  }
   const update = args.input
   return Product.findByIdAndUpdate(args.id, update, { new: true })
     .lean()
@@ -36,6 +46,9 @@ const updateProduct = (_, args, ctx) => {
 }
 
 const removeProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError("you don't have access")
+  }
   return Product.findByIdAndRemove(args.id)
     .lean()
     .exec()
